requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
streamlit>=1.28.0
tavily-python>=0.3.0
pypdf>=3.17.0
python-dotenv>=1.0.0
openai>=1.0.0

